A1 float converted to double, added virtual destructor to class shape

+++ marashov.alexander/A1/base-types.hpp
@@ -3,12 +3,12 @@
 
 struct point_t
 {
-  float x, y;
+  double x, y;
 };
 
 struct rectangle_t
 {
-  float width, height;
+  double width, height;
   point_t pos;
 };
