+++ marashov.alexander/A1/circle.hpp
@@ -6,18 +6,18 @@
 class Circle : public Shape
 {
   public:
-    Circle(const point_t& center, float radius);
-    Circle(float centerX, float centerY, float radius);
+    Circle(const point_t& center, double radius);
+    Circle(double centerX, double centerY, double radius);
 
-    float getArea() const override;
+    double getArea() const override;
     rectangle_t getFrameRect() const override;
     point_t getCenter() const override;
     void move(const point_t& pos) override;
-    void move(float dltX, float dltY) override;
+    void move(double dltX, double dltY) override;
 
   private:
     point_t m_center;
-    float m_radius;
+    double m_radius;
 };
 
 #endif //CIRCLE_HPP
